requirements.txt
fastapi
uvicorn[standard]
playwright
httpx
pandas
pdfplumber  
matplotlib
python-dotenv